Re-organize poke tests, add alloc lints, thanks @epage for the hint
