+++ facet-core/src/_trait/impls/vec_impl.rs
@@ -1,8 +1,10 @@
+extern crate alloc;
+
 use crate::*;
 use core::{alloc::Layout, hash::Hash as _};
 
 #[cfg(feature = "std")]
-use std::vec::Vec;
+use alloc::vec::Vec;
 
 unsafe impl<T> Facet for Vec<T>
 where
