+++ facet-peek/src/lib.rs
@@ -1,4 +1,6 @@
 #![warn(missing_docs)]
+#![warn(clippy::std_instead_of_core)]
+#![warn(clippy::std_instead_of_alloc)]
 #![doc = include_str!("../README.md")]
 
 //! Allows peeking (reading from) shapes
